prints the size of various types on the computer

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long int c;
 long long int d;
 float e;
 
-printf("size of a char: %d byte(s)\n", sizeof(a));
-printf("size of an int: %d byte(s)\n", sizeof(b));
-printf("size of a long int: %d byte(s)\n", sizeof(c));
-printf("size of a long long int: %d byte(s)\n", sizeof(d));
-printf("size of a float: %d byte(s)\n", sizeof(e));
+printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(a));
+printf("size of an int: %d byte(s)\n", (unsigned long)sizeof(b));
+printf("size of a long int: %d byte(s)\n", (unsigned long)sizeof(c));
+printf("size of a long long int: %d byte(s)\n", (unsigned long)sizeof(d));
+printf("size of a float: %d byte(s)\n", (unsigned long)sizeof(e));
 return (0);
 }
